+++ Dragoon Modifier/MemoryController/DragoonLevel.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Dragoon_Modifier.MemoryController {
     public class DragoonLevel {
-        long _baseAddress;
+        int _baseAddress;
 
         public ushort MaxMP { get { return Emulator.ReadUShort(_baseAddress); } set { Emulator.WriteUShort(_baseAddress, value); } }
         public byte SpellLearned { get { return Emulator.ReadByte(_baseAddress + 0x2); } set { Emulator.WriteByte(_baseAddress + 0x2, value); } }
@@ -15,7 +15,7 @@ namespace Dragoon_Modifier.MemoryController {
         public byte DDF { get { return Emulator.ReadByte(_baseAddress + 0x6); } set { Emulator.WriteByte(_baseAddress + 0x6, value); } }
         public byte DMDF { get { return Emulator.ReadByte(_baseAddress + 0x7); } set { Emulator.WriteByte(_baseAddress + 0x7, value); } }
 
-        public DragoonLevel(long baseAddress, int level) {
+        public DragoonLevel(int baseAddress, int level) {
             _baseAddress = baseAddress + level * 0x8;
         }
     }
