+++ Dragoon Modifier/MemoryController/IntCollection.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 
 namespace Dragoon_Modifier.MemoryController {
     public class IntCollection {
-        long _baseAddr;
+        int _baseAddr;
         int _offset;
         uint _size;
 
         public uint Length { get { return _size; } }
 
-        public IntCollection(long baseAddress, int offset, uint numberOfElements) {
+        public IntCollection(int baseAddress, int offset, uint numberOfElements) {
             _baseAddr = baseAddress;
             _offset = offset;
             _size = numberOfElements;
